[WOOCOU-29] ♻️ refactor : pr 피드백 코멘트 반영
