+++ src/main/java/com/coumin/woowahancoupons/coupon/service/CouponService.java
@@ -1,9 +1,8 @@
 package com.coumin.woowahancoupons.coupon.service;
 
 import com.coumin.woowahancoupons.coupon.dto.StoreCouponSaveRequestDto;
-import java.util.List;
 
 public interface CouponService {
 
-    void saveAllStoreCoupons(Long storeId, List<StoreCouponSaveRequestDto> requestDtoList);
+    void saveAllStoreCoupons(Long storeId, StoreCouponSaveRequestDto requestDto);
 }
